requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
flask-cors==4.0.0
psycopg2-binary==2.9.7
pandas==2.1.1
PyPDF2==3.0.1
python-docx==1.1.0
# textract==1.6.4
Werkzeug==2.3.7
openpyxl==3.1.2
alembic==1.12.1